Separate out POST integration tests for relationships
