+++ test/integration/post-resource/success.js
@@ -5,7 +5,6 @@ const getDb = require('../../../lib/database');
 const wipeDatabase = require('../../../lib/database/wipe');
 const validators = require('../../helpers/json-api-validators');
 const applyMigrations = require('../../helpers/apply-migrations');
-const seed = require('../../helpers/seed');
 
 const db = getDb();
 const fixturesDirectory = path.join(__dirname, '..', '..', 'fixtures');
@@ -112,144 +111,6 @@ describe('Resource POST success', function() {
             .expect(201)
             .end(done);
         });
-    });
-  });
-
-  describe('when the request is valid, with a many-to-one relationship', () => {
-    beforeEach((done) => {
-      this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
-        apiVersion: 10
-      };
-
-      const relationGuestSeeds = [{
-        first_name: 'sandwiches'
-      }];
-
-      applyMigrations(this.options)
-        .then(() => seed('relation_guest', relationGuestSeeds))
-        .then(() => done());
-    });
-
-    it('should return a 200 OK, with the created resource', (done) => {
-      const expectedData = {
-        type: 'relations',
-        id: '1',
-        attributes: {
-          name: 'please',
-          size: null
-        },
-        relationships: {
-          owner: {
-            data: {
-              id: '1',
-              type: 'relation_guests'
-            },
-            links: {
-              self: '/v10/relations/1/relationships/owner',
-              related: '/v10/relations/1/owner'
-            }
-          }
-        }
-      };
-
-      const expectedLinks = {
-        self: '/v10/relations/1'
-      };
-
-      request(app(this.options))
-        .post('/v10/relations')
-        .send({
-          data: {
-            type: 'relations',
-            attributes: {
-              name: 'please'
-            },
-            relationships: {
-              owner: {
-                data: {
-                  id: '1',
-                  type: 'relation_guests'
-                }
-              }
-            }
-          }
-        })
-        .expect(validators.basicValidation)
-        .expect(validators.assertData(expectedData))
-        .expect(validators.assertLinks(expectedLinks))
-        .expect('Location', '/v10/relations/1')
-        .expect(201)
-        .end(done);
-    });
-  });
-
-  describe('when the request is valid, with a one-to-one relationship', () => {
-    beforeEach((done) => {
-      this.options = {
-        resourcesDirectory: path.join(fixturesDirectory, 'kitchen-sink'),
-        apiVersion: 10
-      };
-
-      const relationGuestSeeds = [{
-        first_name: 'sandwiches'
-      }];
-
-      applyMigrations(this.options)
-        .then(() => seed('relation_guest', relationGuestSeeds))
-        .then(() => done());
-    });
-
-    it('should return a 200 OK, with the created resource', (done) => {
-      const expectedData = {
-        type: 'one_to_ones',
-        id: '1',
-        attributes: {
-          name: 'please',
-          size: null
-        },
-        relationships: {
-          owner: {
-            data: {
-              id: '1',
-              type: 'relation_guests'
-            },
-            links: {
-              self: '/v10/one_to_ones/1/relationships/owner',
-              related: '/v10/one_to_ones/1/owner'
-            }
-          }
-        }
-      };
-
-      const expectedLinks = {
-        self: '/v10/one_to_ones/1'
-      };
-
-      request(app(this.options))
-        .post('/v10/one_to_ones')
-        .send({
-          data: {
-            type: 'one_to_ones',
-            attributes: {
-              name: 'please'
-            },
-            relationships: {
-              owner: {
-                data: {
-                  id: '1',
-                  type: 'relation_guests'
-                }
-              }
-            }
-          }
-        })
-        .expect(validators.basicValidation)
-        .expect(validators.assertData(expectedData))
-        .expect(validators.assertLinks(expectedLinks))
-        .expect('Location', '/v10/one_to_ones/1')
-        .expect(201)
-        .end(done);
     });
   });
